Added auto-updating of GoFormation resources
Added a travis script that checks on a daily basis for any AWS
CloudFormation updates (comparing against master branch), and if any are
found, create a new pull request for them automatically.

If more changes come in before the pull request is closed/merged, the
commits will be added to the pull request.