do not omit `ReturnData`
AWS documentation says 

*ReturnData*
This option indicates whether to return the timestamps and raw data values of this metric. If you are performing this call just to do math expressions and do not also need the raw data returned, you can specify False. *If you omit this, the default of True is used.*

Then, cloudformation template update with fail with the error: `Exactly one element of the metrics list should return data`, because  all the omitted values assumed as `True`