Workaround for AWS::SSM::Association.Parameters
On 2021-04-28, the CFN spec for AWS::SSM::Association.Parameters
specified a (potentially invalid) property type that included Type=Map,
and ItemType=List, without any further type information on what the list
should consist of. If this happens, all we can do is assume it's a list
of strings. Aka, map[string][]string.