BREAKING CHANGE: Removed previous globals impl
This change removes the previous globals implementation. Previously when
parsing a JSON/YAML template that contained globals, the parser would
automatically overwrite the properties in the parsed resources with the
value from the global defined (essentially what CFN does on the
backend).

With the changes in PR #376, this is superceded with a new globals
implementation that allows full composition and decomposition of
templates that contain a global section - WITHOUT overwriting the
values.