chore(tests): Fixed tests that relied on required properties
We had some tests that relied on the fact that AWS::Route53::HostedZone
had a required property of 'Name'. This recently changed and the Name
property is no longer marked as required in the CFN schema, so tests
failed.

Searched for resources that had a single required property and picked
one at random (AWS::IoT::Policy). Tests passing now.