fix(schema): AWS::CDK::Metadata should be generated
In PR #418, the AWS::CDK::Metadata was kindly added, however the
resource was manually created, which lead to it being overwritten when
`go generate` was run to update the goformation schema from
CloudFormation/SAM specifications.

This change adds a new `generate/cdk.json` file that contains a manually
crafted version of the AWS::CDK::* specification (as I don't think an
automatically published one exists), and automatically generates the
`cdk/aws-cdk-metadata.go` resource from it.

One side effect of this is that the `AWS::CDK::Metadata` resource is no
longer in the main `cloudformation.schema.json` that some IDEs use for
autocompletion/intellisense, and it now lives in an autogenerated
`schema/cdk.schema.json` file (similar to how SAM is separate, as it's
generated from a separate resource definition file).